refactor: remove some data fetching from atom tree
This commit removes some of the atoms that fetch data from the API.
The motivation for this came when attempting to upgrade Jotai, and
running into multiple issues doing so, esp. with the query toolkit.
As the objective is to remove dependence on this, I had a crack at
removing some, which is extremely difficiult given the inter-conn-
mesh nature of our atom set up.