refactor: initial app route, remove legacy prop
This change removes the `hasSetPassword` value, as this was a legacy
implementation detail when a wallet could have a pre-, and post-,
having set a password states. As we now only "create" a wallet when
setting the password, this state can be inferred, and thus removed.
Further, this allows us to simplify the account guard logic.